// app/commons/enum.ts
export enum APP_ENVIRONMENT {
	LOCAL = 'local',
	DEV = 'dev',
	PRODUCTION = 'production',
}
export enum ENV {
	PORT = 'PORT',
	VERSION = 'VERSION',
	APP_ENV = 'APP_ENV',
	MORGAN_ENV = 'MORGAN_ENV',
	/* Database MySQL */
	DATABASE_MYSQL_HOST = 'DATABASE_MYSQL_HOST',
	DATABASE_MYSQL_USER = 'DATABASE_MYSQL_USER',
	DATABASE_MYSQL_PASSWORD = 'DATABASE_MYSQL_PASSWORD',
	DATABASE_MYSQL_DATABASE = 'DATABASE_MYSQL_DATABASE',
	DATABASE_MYSQL_CONNECTION_LIMIT = 'DATABASE_MYSQL_CONNECTION_LIMIT',
	/* Database PostgreSQL */
	DATABASE_POSTGRESQL_HOST = 'DATABASE_POSTGRESQL_HOST',
	DATABASE_POSTGRESQL_USER = 'DATABASE_POSTGRESQL_USER',
	DATABASE_POSTGRESQL_PASSWORD = 'DATABASE_POSTGRESQL_PASSWORD',
	DATABASE_POSTGRESQL_DATABASE = 'DATABASE_POSTGRESQL_DATABASE',
	DATABASE_POSTGRESQL_CONNECTION_LIMIT = 'DATABASE_POSTGRESQL_CONNECTION_LIMIT',
	/* Database MongoDb */
	DATABASE_MONGODB_HOST = 'DATABASE_MONGODB_HOST',
	DATABASE_MONGODB_URI = 'DATABASE_MONGODB_URI',
}
export enum HEADERS {
	ACCESS_CONTROL_ALLOW_HEADERS = 'Access-Control-Allow-Headers',
	ACCESS_CONTROL_ALLOW_ORIGIN = 'Access-Control-Allow-Origin',
	ACCESS_CONTROL_ALLOW_METHODS = 'Access-Control-Allow-Methods',
	X_AMZ_DATE = 'X-Amz-Date',
	AUTHORIZATION = 'Authorization',
	X_API_KEY = 'X-Api-Key',
	X_AMZ_SECURITY_TOKEN = 'X-Amz-Security-Token',
	CONTENT_TYPE = 'Content-Type',
	CONTENT_TYPE_LOWER = 'content-type',
}
export enum HEADERS_VALUE {
	ALLOW = '*',
	APPLICATION_JSON = 'application/json',
}
export enum HTTP {
	STATUS_CODE_201 = 201,
	STATUS_CODE_200 = 200,
	STATUS_CODE_400 = 400,
	STATUS_CODE_401 = 401,
	STATUS_CODE_403 = 403,
	STATUS_CODE_404 = 404,
	STATUS_CODE_500 = 500,
}
export enum TAG {
	ERROR_UTIL = 'Util ErrorUtil             | ',
	CONTROLLER_PERSON = 'PersonController           | ',
}
export enum NAME {
	ERROR_UTIL = 'errorUtil()',
}
export enum LOG_TYPE {
	INFO = 'info',
	TRACE = 'trace',
	DEBUG = 'debug',
	FATAL = 'fatal',
	WARN = 'warn',
}
export enum LOG_MODE {
	SHORT = 'short',
	LONG = 'long',
	SIMPLE = 'simple',
	JSON = 'json',
	BUNYAN = 'bunyan',
	INSPECT = 'inspect',
}
export enum NUM {
	ZERO = 0,
	ONE = 1,
	TWO = 2,
	THREE = 3,
	FOUR = 4,
	FIVE = 5,
	SIX = 6,
	SEVEN = 7,
	EIGHT = 8,
	NINE = 9,
	TEN = 10,
}
export const NUMERIC_BASE = {
	BINARY: 2, // Base binaria (0-1)
	OCTAL: 8, // Base octal (0-7)
	DECIMAL: 10, // Base decimal (0-9)
	HEXADECIMAL: 16, // Base hexadecimal (0-F)
};